backend_saas/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
mysql-connector-python
python-jose
passlib[bcrypt]
pydantic
python-dotenv 